修正腳色無法向左右轉向的bug

+++ game4.10/Source/persona.h
@@ -17,10 +17,10 @@ class persona : public BasicObject
 		void SetReloading(bool flag);
         void setCan_move(bool flag);
         void setBullet(int num);
-		void setFacingPosition(int,int);
+		void setFacingPosition(double,double);
 		void setDirection();
-		int getFacingX();
-		int getFacingY();
+		double getFacingX();
+		double getFacingY();
 		int getHasitemNum();
 		int getBullet();
         bool isGetting();
@@ -35,7 +35,7 @@ class persona : public BasicObject
 		bool is_Reloading = false;
         vector<items> hasitem;
         int bullet;
-		int facingX,facingY;
+		double facingX,facingY;
 		int direction;
 
 };
